fix[coming-soon]: redirect
